pyspark-transactions/requirements.txt
# Runtime — Spark 4.0
# Requires: Java 17 or 21, Python 3.9+
pyspark>=4.0.0,<4.1.0
requests>=2.31
pyyaml>=6.0

# Orchestration
apache-airflow>=2.8,<3.0
apache-airflow-providers-apache-spark>=4.0

# Data quality
great-expectations>=1.0,<2.0
pandas>=2.0
